Updated the Login, Subscription, Payment Method, Mainpage, and Auth
Making the login, registration, subscription, and payment-method functionable. Dealt with errors such as not being able to login even when registered, unable to access the other htmls once proceeding to sign in button from index.html, to store information from the database and then check for duplications, not allowing same mobile number input for Mobile and Digital, double checking html for not loading their css and scripts, and had some improvement and tweaking to index to possible make it work.